requirements.txt
biopython==1.86
pandas==3.0.0
numpy==2.4.2
scipy==1.17.0
scikit-learn==1.8.0
fastapi==0.128.2
uvicorn==0.40.0
pytest==9.0.2
